# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  emas_internal_handler_dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  handler.SettingsResponse:
    properties:
      ai_enabled:
        type: boolean
      integrations:
        items:
          type: string
        type: array
      language:
        type: string
      notifications:
        type: boolean
      theme:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /scheduling/backfill-training-dataset:
    get:
      consumes:
      - application/json
      description: Backfill training dataset
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Backfill training dataset
      tags:
      - scheduling
  /scheduling/candidate-machines:
    get:
      consumes:
      - application/json
      description: Candidate machines
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Candidate machines
      tags:
      - scheduling
  /scheduling/estimate-job-completion:
    get:
      consumes:
      - application/json
      description: Estimate job completion
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Estimate job completion
      tags:
      - scheduling
...
  /scheduling/explosion:
    get:
      consumes:
      - application/json
      description: Explode demand
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Explode demand
      tags:
      - scheduling
...
  /scheduling/readiness:
    get:
      consumes:
      - application/json
      description: Check readiness
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Check readiness
      tags:
      - scheduling
  /scheduling/refresh-work-calendars:
    get:
      consumes:
      - application/json
      description: Refresh work calendars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Refresh work calendars
      tags:
      - scheduling
...
  /scheduling/solver-preview:
    get:
      consumes:
      - application/json
      description: Solver preview
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Solver preview
      tags:
      - scheduling
  /scheduling/training-dataset:
    get:
      consumes:
      - application/json
      description: Export training dataset
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Export training dataset
      tags:
      - scheduling
  /scheduling/training-dataset-stats:
    get:
      consumes:
      - application/json
      description: Training dataset stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Training dataset stats
      tags:
      - scheduling
  /scheduling/validate-slot:
    post:
      consumes:
      - application/json
      description: Validate slot
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Validate slot
      tags:
      - scheduling
...
  /settings/get:
    get:
      consumes:
      - application/json
      description: Get settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Get settings
      tags:
      - settings
  /settings/update:
    put:
      consumes:
      - application/json
      description: Update settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/emas_internal_handler_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.SettingsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emas_internal_handler_dto.Response'
      summary: Update settings
      tags:
      - settings
...
swagger: "2.0"